Applicate modifiche tutorial
