+++ Assets/Scripts/TutorialScript.cs
@@ -6,25 +6,29 @@ public class TutorialScript : MonoBehaviour
 {
     public GameObject[] istruzioni;
     public LogicScript logic;
-    int i = 0;
-    // Update is called once per frame
+    public int i = 0;    
     public void Update()
     {
-        if (Input.GetAxisRaw("Horizontal")== 1 && i < 3)
+        if (Input.GetButtonDown("Right") && i < 3)
         {
-                istruzioni[i].SetActive(false);
-                i++;
-                istruzioni[i].SetActive(true);
+            right();
         }
-        if (Input.GetAxisRaw("Horizontal") == -1 && i > 0)
+        if (Input.GetButtonDown("Left") && i > 0)
         {
-            istruzioni[i].SetActive(false);
-            i--;
-            istruzioni[i].SetActive(true);
+            left();
         }
-        if (Input.GetButtonDown("FireRed"))
-        {
-            logic.HowtoPlayExitButton();
-        }
+    }
+    public void right()
+    {
+        istruzioni[i].SetActive(false);
+        i++;
+        istruzioni[i].SetActive(true);
+    }
+    public void left()
+    {
+        istruzioni[i].SetActive(false);
+        i--;
+        istruzioni[i].SetActive(true);
+
     }
 }
